+++ lib/friendly_id/active_record.rb
@@ -1,22 +1,13 @@
 module FriendlyId
 
-  # Are we running on ActiveRecord 3 or higher?
-  def self.on_ar3?
-    ActiveRecord::VERSION::STRING >= "3"
-  end
-
   module ActiveRecordAdapter
 
     include FriendlyId::Base
 
     def has_friendly_id(method, options = {})
-      if FriendlyId.on_ar3?
-        class_attribute :friendly_id_config
-        self.friendly_id_config = Configuration.new(self, method, options)
-      else
-        class_inheritable_accessor :friendly_id_config
-        write_inheritable_attribute :friendly_id_config, Configuration.new(self, method, options)
-      end
+
+      class_attribute :friendly_id_config
+      self.friendly_id_config = Configuration.new(self, method, options)
 
       if friendly_id_config.use_slug?
         include SluggedModel
@@ -47,7 +38,6 @@ end
 
 require "friendly_id/active_record_adapter/relation"
 require "friendly_id/active_record_adapter/configuration"
-require "friendly_id/active_record_adapter/finders"
 require "friendly_id/active_record_adapter/simple_model"
 require "friendly_id/active_record_adapter/slugged_model"
 require "friendly_id/active_record_adapter/slug"
@@ -56,18 +46,11 @@ require "friendly_id/active_record_adapter/tasks"
 module ActiveRecord
   class Base
     extend FriendlyId::ActiveRecordAdapter
-    unless FriendlyId.on_ar3?
-      class << self
-        VALID_FIND_OPTIONS << :scope
-      end
-    end
   end
 
-  if defined? Relation
-    class Relation
-      alias find_one_without_friendly  find_one
-      alias find_some_without_friendly find_some
-      include FriendlyId::ActiveRecordAdapter::Relation
-    end
+  class Relation
+    alias find_one_without_friendly  find_one
+    alias find_some_without_friendly find_some
+    include FriendlyId::ActiveRecordAdapter::Relation
   end
 end