+++ lib/friendly_id/active_record_adapter/simple_model.rb
@@ -10,7 +10,6 @@ module FriendlyId
           validates_presence_of column, :unless => :skip_friendly_id_validations
           validates_length_of column, :maximum => friendly_id_config.max_length, :unless => :skip_friendly_id_validations
           after_update :update_scopes
-          extend FriendlyId::ActiveRecordAdapter::Finders unless FriendlyId.on_ar3?
         end
       end