+++ test/active_record_adapter/ar_test_helper.rb
@@ -108,8 +108,7 @@ end
 # A model that uses default slug settings and has a named scope
 class Post < ActiveRecord::Base
   has_friendly_id :name, :use_slug => true
-  def self.named_scope(*args) scope(*args) end if FriendlyId.on_ar3?
-  named_scope :published, :conditions => { :published => true }
+  scope :published, :conditions => { :published => true }
 end
 
 # Model that uses a custom table name